requirements.txt
cffi==1.15.1
cryptography==39.0.1
paho-mqtt==1.6.1
pycparser==2.21
